fix : Redis에서 값을 가져올 때 생기는 에러 해결
Redis에서 ValueOperations 가져올 때 값을 Long으로 가져오지 못해 에러가 발생해 Integer로 type을 바꿔 해결함

+++ user-api/src/main/java/zerobase/bud/chat/service/ChatRoomService.java
@@ -53,7 +53,7 @@ public class ChatRoomService {
     @Transactional(readOnly = true)
     public Slice<ChatRoomDto> searchChatRooms(String keyword, int page) {
 
-        ValueOperations<String, Long> valueOperations = redisTemplate.opsForValue();
+        ValueOperations<String, Integer> valueOperations = redisTemplate.opsForValue();
 
         return chatRoomRepository
                 .findAllByTitleContainingIgnoreCaseAndStatus(keyword, ACTIVE,
@@ -66,7 +66,7 @@ public class ChatRoomService {
     @Transactional(readOnly = true)
     public Slice<ChatRoomDto> readChatRooms(int page) {
 
-        ValueOperations<String, Long> valueOperations = redisTemplate.opsForValue();
+        ValueOperations<String, Integer> valueOperations = redisTemplate.opsForValue();
 
         return chatRoomRepository.findAllByStatus(ACTIVE,
                         PageRequest.of(page, CHATROOM_SIZE_PER_PAGE))
@@ -78,7 +78,7 @@ public class ChatRoomService {
     @Transactional(readOnly = true)
     public ChatRoomDto readChatRoom(Long chatroomId) {
 
-        ValueOperations<String, Long> valueOperations = redisTemplate.opsForValue();
+        ValueOperations<String, Integer> valueOperations = redisTemplate.opsForValue();
 
         return ChatRoomDto.of(
                 chatRoomRepository.findByIdAndStatus(chatroomId, ACTIVE)
@@ -87,8 +87,8 @@ public class ChatRoomService {
         );
     }
 
-    private Long getNumberOfMembers(Long chatroomId, ValueOperations<String, Long> valueOperations) {
-        return Optional.ofNullable(valueOperations.get(CHATROOM + chatroomId)).orElse(1L);
+    private Integer getNumberOfMembers(Long chatroomId, ValueOperations<String, Integer> valueOperations) {
+        return Optional.ofNullable(valueOperations.get(CHATROOM + chatroomId)).orElse(1);
     }
 
     @Transactional(readOnly = true)
